Fix of issue #101 in https://github.com/Jaguar-dart/jaguar_orm/issues/101:
In case of ManyToMany relation, other side of relation will always be deleted when cascading removal, even if another relation exists.

In addition to this, there may be use cases in which you don't want to remove orphaned entities. Because of this, a new optional bool parameter orphanRemoval has been added to methods remove and detachXXX.